// ProcessSIM/ProcessSIM.Application/ClientApp/src/app/models/complexity.interface.ts
export interface IComplexity {
  originalDetailsCountValue: number;
  isStandardDetailsValue: number;
  isAssemblyUnitsSurfValue: number;
  isAssemblyUnitsConnValue: number;
  surfaceFormValue: number;
  isStructuralElementsValue: number;
  workWithPartsValue: number;
  isRestrictDetailsValue: number;
}